requirements.txt
alabaster==1.0.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
django-cors-headers==4.4.0
django-eventstream==5.3.1
django-grip==3.5.1
django-guest-user==0.5.5
djangorestframework==3.15.2
docutils==0.21.2
gripcontrol==4.2.0
hyperlink==21.0.0
idna==3.10
imagesize==1.4.1
incremental==24.7.2
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pubcontrol==3.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
requests==2.32.3
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinx-rtd-theme==3.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
zope.interface==7.0.3
